requirements.txt
# ============================================================================
# DEPENDENCIAS PYTHON - DATA WAREHOUSE ETL
# ============================================================================

# Conexión a Bases de Datos
# -----------------------------------------------------------------------------
# MariaDB/MySQL (PyMySQL es puro Python, no requiere compiladores)
PyMySQL==1.1.0

# SQL Server (pyodbc requiere ODBC Driver 17 for SQL Server)
pyodbc==5.0.1

# Manipulación de Datos
# -----------------------------------------------------------------------------
pandas==2.1.3
numpy==1.26.2

# Configuración y Variables de Entorno
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging Avanzado
# -----------------------------------------------------------------------------
loguru==0.7.2

# Fechas y Tiempo
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
pytz==2023.3

# Validación de Datos
# -----------------------------------------------------------------------------
pydantic==2.5.2
great-expectations==0.18.7

# Utilidades
# -----------------------------------------------------------------------------
tqdm==4.66.1  # Barras de progreso
colorama==0.4.6  # Colores en consola (Windows)

# Testing
# -----------------------------------------------------------------------------
pytest==7.4.3
pytest-cov==4.1.0

# Monitoreo y Performance
# -----------------------------------------------------------------------------
psutil==5.9.6  # Monitoreo de recursos del sistema

# Notificaciones
# -----------------------------------------------------------------------------
# Descomenta si necesitas notificaciones por email
# yagmail==0.15.293

# Power BI
# -----------------------------------------------------------------------------
# Descomenta si necesitas conexión directa con Power BI Service
# msal==1.25.0
# azure-identity==1.15.0

# Desarrollo
# -----------------------------------------------------------------------------
ipython==8.18.1  # Consola interactiva mejorada
black==23.11.0  # Formateador de código
flake8==6.1.0  # Linter

# Opcional: Para procesamiento paralelo más avanzado
# -----------------------------------------------------------------------------
# joblib==1.3.2
# dask==2023.11.0
